update to build
git-svn-id: https://zoop.googlecode.com/svn/trunk@378 585e9b84-f019-0410-b165-1d737e75257b

+++ framework/build/functions.php
@@ -86,6 +86,8 @@ function _gen($path, $filePath = '', $params = array())
 	
 	$content = $gui->fetch($templatePath . '.tpl');
 	_status("creating generated file '" . getcwd() . '/' . $filePath . "'");
+	if(file_exists($filePath))
+		die("file $filePath already exists\n");
 	file_put_contents($filePath, $content);
 }
